wolf-memory/requirements.txt
ollama
mcp
requests
PyYAML
